backend/proxy/requirements.txt
Flask==2.1.1
requests==2.26.0
google-auth==2.0.2
